extracted quiz funcitonality to new file
